+++ apps/vmq_diversity/test/vmq_diversity_plugin_SUITE.erl
@@ -47,6 +47,7 @@ all() ->
      on_client_wakeup_test,
      on_client_offline_test,
      on_client_gone_test,
+     on_session_expired_test,
      auth_on_register_undefined_creds_test,
      invalid_modifiers_test,
 
@@ -126,7 +127,7 @@ on_unsubscribe_test(_) ->
 
 on_deliver_test(_) ->
     ok = vmq_plugin:all_till_ok(on_deliver,
-                                [username(), allowed_subscriber_id(), topic(), payload()]).
+                                [username(), allowed_subscriber_id(), 1, topic(), payload(), false]).
 
 on_offline_message_test(_) ->
     [next] = vmq_plugin:all(on_offline_message, [allowed_subscriber_id(), 2,
@@ -137,6 +138,8 @@ on_client_offline_test(_) ->
     [next] = vmq_plugin:all(on_client_offline, [allowed_subscriber_id()]).
 on_client_gone_test(_) ->
     [next] = vmq_plugin:all(on_client_gone, [allowed_subscriber_id()]).
+on_session_expired_test(_) ->
+    [next] = vmq_plugin:all(on_session_expired, [allowed_subscriber_id()]).
 
 auth_on_register_undefined_creds_test(_) ->
     Username = undefined,
@@ -228,7 +231,7 @@ on_publish_m5_test(_) ->
     [next] = vmq_plugin:all(on_publish_m5, Args).
 
 on_deliver_m5_test(_) ->
-    Args = [username(), allowed_subscriber_id(), topic(), payload(),
+    Args = [username(), allowed_subscriber_id(), 1, topic(), payload(), false,
             #{?P_USER_PROPERTY =>
                   [{<<"k1">>, <<"v1">>},
                    {<<"k2">>, <<"v2">>}],